pyproject.toml
[tool.poetry]
name = "gft-usher-tools"
version = "0.1.0"
description = ""
authors = ["evangriffiths <egriffiths93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gspread = "^6.0.2"
pandas = "^2.2.0"
python-dotenv = "^1.0.1"
oauth2client = "^4.1.3"
openpyxl = "^3.1.2"
pydantic = "^2.6.1"
google-api-python-client = "^2.118.0"
pytest = "^8.0.0"
pygsheets = "^2.0.6"
bs4 = "^0.0.2"
selenium = "^4.17.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
